fixes bug with logging an error response
trying to merge the resopnse object into the logged object with the spread operator `{...res,` resulted in this error:

```
TypeError: One of the sources for assign has an enumerable key on the prototype chain. Are you trying to assign a prototype property? We don't allow it, as this is an edge case that we do not support. This error is a performance optimization and not spec compliant.
```